requirements.txt
astronomer-cosmos>=1.0.2
apache-airflow-providers-amazon==7.4.1
sqlparse==0.4.2
pytest>=7.0.0
pytest-cov>=4.0.0
coverage>=7.0.0
pre-commit
